refactor(vite-plugin-angular): rename utils/shared.ts to utils/compilation-shared.ts
