+++ packages/vite-plugin-angular/src/lib/utils/compilation-shared.ts
